2008-10-25  Geoffrey Garen  <ggaren@apple.com>
        Not reviewed.

        Try to fix Windows build.

        * VM/SamplingTool.h:
        (JSC::SamplingTool::encodeSample): Explicitly cast bool to int, to
        silence compiler warning.



Canonical link: https://commits.webkit.org/30115@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@37893 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ JavaScriptCore/VM/SamplingTool.h
@@ -160,7 +160,7 @@ namespace JSC {
         unsigned encodeSample(Instruction* vPC, bool inCTIFunction = false, bool inHostFunction = false)
         {
             ASSERT(!(reinterpret_cast<intptr_t>(vPC) & 0x3));
-            return reinterpret_cast<intptr_t>(vPC) | (inCTIFunction << 1) | inHostFunction;
+            return reinterpret_cast<intptr_t>(vPC) | (static_cast<intptr_t>(inCTIFunction) << 1) | static_cast<intptr_t>(inHostFunction);
         }
 
     private:
